SchemaToProto: Move common utils out of jtd_to_proto.py
